requirements.txt
# pip install -r requirements.txt

numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0
ccxt>=4.0.0
yfinance>=0.2.0
pandas_ta>=0.3.14b
PyPortfolioOpt>=1.5.0
# pypfopt>=1.5.0 # This byd pypfopt does not fucking work on the bloody Python 3.14. This is the same as PyPortfolioOpt. The fork has some bug fixes and improvements.
pyyaml>=5.4.0
statsmodels>=0.13.0
pyarrow>=10.0.0